endgame: endgame_t 由8位长度改为32位因对缩小残局库文件体积无帮助
为调试方便，修改长度。

+++ src/ai/endgame.h
@@ -32,7 +32,8 @@
 using namespace std;
 using namespace CTSL;
 
-enum endgame_t : uint8_t
+// TODO: uint8_t
+enum endgame_t : uint32_t
 {
     ENDGAME_NONE,
     ENDGAME_PLAYER_BLACK_WIN,
